Integrated imputation into Stochastic class.
Data Stochastics will now automatically impute missing data. This behavior
will be triggered when a numpy MaskedArray is passed as the value and the
observed flag is set to True. This was done with very little mangling of
the Stochastic code, so hopefully it is not too disruptive.

Further testing needed before merging with master.