Install git version of Theano
The current release of Theano on PyPI only supports python 2. This
change install from the git master branch instead (for both python 2 and
python 3).

Using '<=0.6.1dev' as the required version is a hack to get pip to
install the git version of Theano. If set to '==0.6.1dev' pip will
download the git version but then throw an error because the git version
is still 0.6.0rc3. If set to '==0.6.0rc3', pip will give preference to
the pypi version.

requirements.txt, which is no longer read into setup.py, has been
removed.

Fixes #388