Better traceplot for discrete variables
This provides better histograms for traceplotting discrete variables. The old
version set the number of bins=sqrt(len(trace)), this meant smaller and smaller
vertical bars for long traces. Now it calculates the number of bins based on
the minimum and maximum value in the trace, centering bars at the integer
values present. This assumes discrete distributions only take on integer
values, though this might not always be true.