Add base and NDArray backend
This commit contains a new backend for sampling and selecting values.
Non-backend files have been changed to work with the new backend.
Everything seems to be working with the exception of two issues (marked
with FIXME):

1. pymc.plots.forestplot has not been updated yet for the new backend.
2. The previous behavior of passing a trace object to sample is not the
   same. I updated stochastic_volatility to do this with the same trace
   object.

This commit also introduces a change to `sample`/`psample`. Instead of
having separate function, `sample` now takes a keyword argument
`threads`, and if this is over one, the multiprocessing version is used.

The method for selecting values has also been changed. Traces can still
be indexed to return values, a new slice, or a point (depending on the
index), but the handling of chains is different. The trace object is now
manages multiple chains itself instead of having a separate class to
manage the single trace object. `get_values` is the main method for
selecting values. By default, it returns separate results for all the
chains. The chains can be combine with the `combine` flags, and
particular chains can be select with the `chains` argument.

The motivation for both sample and selection changes above was to have a
unified interface for dealing with multiple chains, as most people are
likely going to take advantage of the parallel sampling.