Revert "Rename sample.py to sampling.py"
This reverts commit 683507b08504c751a32dcc7e12bc2711bfbe2ecc.

The is being reverted to make it easier to compare changes with the
master branch.

However, this reintroduces the naming conflict that the commit fixed.
'pymc.sample' now refers to the function, but the module of the same
name cannot be accessed. This makes it difficult to test anything other
than the top-level functions that are imported into pymc/__init__.py
(sample and iter_sample).

Conflicts:
        pymc/__init__.py