Add summaries for variables over one dimension
Starting with commit 92f2212, `pymc.summary` gave a warning for
variables over 1d (instead of throwing an uncaught NumPy error) because
there wasn't anything in place to deal with displaying the additional
dimensions.

This commit addresses that by grouping the output in the summary table
to reflect the different dimensions. For example, a 2d variable x would
have a 3d MCMC trace, and the summary statistics for x would have a
shape corresponding to the original variable.

    >>> trace['x'].shape
    (3000, 2, 3)
    >>> trace['x'].mean(0).shape
    (2, 3)

To display this in the summary table, the indices for the stats array
are grouped by indices for all dimensions except the rightmost. In the
summary output, the left indices are separated by blank row, while the
rightmost index is shown as rows within a group.

    (0, 0)
    (0, 1)
    (0, 2)

    (1, 0)
    (1, 1)
    (1, 2)

This scheme isn't great visually for variables over 2d because then
there is more than one left index and a blank row is inserted regardless
of which left index is incremented. It'd be possible to extend this by
displaying the current left indices instead of a blank row for variables
over 2d.

Also, this commit changes the naming in the summary tests to refer to
variable dimensions in a consistent way (0d, 1d, 2d, ...). These are
always meant to refer to the original variable dimensions, not the MCMC
trace dimensions.