BUG Fix bounds of Wishart and add warning.
The Wishart has never worked correctly and produced
divergent samples. One of the reasons was that we did
not enforce the matrix to be symmetric nor positive
semi-definite.

Now, however, it is impossible to randomly sample a
valid matrix so we raise a warning that the Wishart
is basically unusable currently and to instead
use the LKJ prior which has many desriable properties.

For more discussions, see https://github.com/pymc-devs/pymc3/issues/538.