tested log(p) of 2D dirichlet
The main goal of this test is to show that the variables are independent
across the second dimension.
Note that betafn and dirichlet_logpdf had to be changed so that
internally they only sum over the first dimension, and only in the final
summation is the log(p) summed over all dimensions.