BUG: ValueError with complete data as MaskedArray
Similar to #529 but a distinct problem: I was passing in a MaskedArray that had no missing values and was getting `ValueError: array is not broadcastable to correct shape`

After bouncing around between pymc3 and Theano for a bit I tracked it down to this: when as_tensor is given a masked array with no mask (data.mask is False), it ends up raising that error. My solution here is to check for that case specifically, but this may not be the most elegant solution.