Add another level to doc folder copying in setup
Package data wasn't copied correctly because of the behaviour of
python's glob module, which needed a further '/*' appended to it.

This affects docs/source/data/6bb8d06c69c0666e6da14c094d4320d115f1ffc8
which is a directory but isn't processed as such by glob, resulting in a
build error.