Suggestion to fix the TypeError issue #994 on Windows.
Errors occurs because some ndarrays with dtype=int32
are passed to the function delta_logp which expects
ndarrays with dtype=int64 as inputs. Inputs are casted
with `ndarray.astype(int)`, but on Windows, the
Python 'int' type is treated as a 32-bit integer
by NumPy.

I suggest to remove the line with `f.trust_input = True`
to ensure that Theano always checks the inputs and
casts it if necessary and when possible.
This fix resolves the issue on my computer:

```
Python version:
2.7.12 |Continuum Analytics, Inc.| (default, Jun 29 2016, 11:07:13) [MSC v.1500 64 bit (AMD64)]
OS version:
Windows-8.1-6.3.9600
```