Implement log CDF for Flat distribution
The reasoning behind this is that a Flat distribution is a Uniform distribution
with bounds stretching symmetrically towards negative and positive infinity.
Or a Normal distribution with infinite standard deviation.

Both of these give values of 1/2 for any finite point in the CDF.
At negative infinity, the CDF is 0 and at positive infinity it is 1.