Fixed step method argument storage hack for unpickling in parallel processes (#2425)
* Fixed step method argument storage hack for unpickling in parallel processes

The original hack did not store the model. As a result, when multiprocessing unpickled the step method, it would attempt to find a model in the context. However there is no context in the forked process, so it fails. This proposed change explicitly stores the model as an argument to be used in constructing the step method object when unpickling it.

* Made the change compatible with python 2.7

* Reverted change