Fix autocorrplot to work with non-sequential chain numbers.
MultiTrace enforces unicity of the chain numbers, but not that they be
sequential (e.g. chains may be [0, 2, 3]). This fix removes the assumption in
autocorrplot that the chains are sequential and start from 0.

As a corollary, the chain number given in the axis titles matches the index of
the underlying chain, rather than just be a number that is incremented for each
chain. This means printed chain numbers start at 0, rather than 1 as they
previously did.