Improve random for ObservedRV
To generate random array for ObservedRV in sample_ppc and sample_prior_predictive a shape kwarg sometimes is needed to make sure the generated array has the same shape as the observed.
This PR improved that to make sure the shape is inferred during random number generation.
Previously I tried to add the shape attr to ObservedRV at the model creation, but turns out it breaks the sample_ppc for theano.shared observation. I try a different approach here to always reset the distribution shape at generation.