Improve documentation of transform aliases.
The transforms code has lower-case alternatives to the upper-cased
class names.  Sometimes these are instances of the classes (e.g.,
`log` is an instance of the `Log` transform), but sometimes they are
aliases (e.g., `lowerbound` is an alias for `LowerBound`).

I couldn't find a good way to document these in the source file, so
they are hand-documented in transforms.rst.