Add Lognormal with location paramter to continous class
After working on a custom lognormal distribution with a location parameter (theta), i was asked to add it too pymc3 through a PR. I have tried to add the location paramter to the lognormal function under the built-in functions. I am not entirely sure if this works completely as intended, there may need to be a tt.switch to check if theta>value, but i have tried to add that condition to bounded in logp.