fix underlying bug behind DensityDist stackoverflow tl:dr: The hashable helper function did not appropiately deal with tuples (and the test case did not actually test the memoization).
In the process of prior-predictive sampling a model involving a DensityDist, the _compile_theano_function function was called with arguments (sd__log_, []).
The _compile_theano_function has a pm.memo.memoize-decorator, which relies on the pm.memo.hashable for hashing of typically unhashable objects.
The "hashable" function incorrectly handled tuples, eventually causing a stackoverflow error on Windows.