Keep broadcasting information in make_shared_replacements
It seems like broadcasting information gets lost when applying
`pm.make_shared_replacements`, leading to problems with the metropolis
sampler. Potentially related issues below:
 - https://github.com/pymc-devs/pymc3/issues/1083
 - https://github.com/pymc-devs/pymc3/issues/1304
 - https://github.com/pymc-devs/pymc3/issues/1983

This fix was previously suggested in the following issue:
 - https://github.com/pymc-devs/pymc3/issues/3337

It could be that further adaptations are necessary as indicated in the
issue. Strangely, this does not seem to lead to problems when using
NUTS.