Break down tests having multiple checks and xfail decorated (#4497)
* Break down tests having multiple checks and xfail decorated

Xfail is going to care only about the first failed assert statement and
ignore all the other ones in the same test.

Sometimes xfail is being used to decorate a test or class function calling
several other sub-functions. This makes it hard to monitor what is happening
for the sub-fuctions that are not failing (given that the failure will take
priority and mark the test function as xfailed). By breaking down each
function into individual sub-functions it's easier to monitor individual
tests behavior

* Remove xfail which seems unnecessary

Also add comment regarding test with unpredictable outcome

* Break down test with multiple sub-checks and xfail where appropriate

* Remove Scipy Xfail from tests not using sp.betabinom

* Add back xfail to potentially failing test when float32 is used

* Temporatily add n_samples=-1 to check tests behavior for all possible parameters

* Remove temporary n_samples and a couple more to do some quick tests

* Remove unnecessary comments

* Replace Xfail with skipif when using Scipy version as condition

Given that betabinom won't exist in previous versions, there should be
no chance for it's behavior to change, given that it won't exist there
is no point in monitoring the outcome of these tests, so it may be worth
skipping them. Skipping them should be faster than running them and
catching the failures

* Add missing n_samples for test_beta_binomial_selfconsistency

* Remove all the n_samples=-1 added for verifying tests robustness

* Tweak normal_logcdf and moyal tests

Temporarily set n_samples=-1 to test all permutations.

* Reintroduce 32bit xfail on Moyal logcdf, revert n_samples and merge passing normal tests

* Small renaming and reordering

* Reintroduce accidentally removed n_samples from binomial test

Co-authored-by: Ricardo <ricardo.vieira1994@gmail.com>