Remove default transformed RV Ops
This functionality was not being used by PyMC. In theory, it could be used to cache one type of transformed RV Op instead of recreating it for every instance of the same RV + transform, resulting in a faster logp rewrite (?). We can always revisit this idea later, but for now it unnecessarily complicated our codebase.