The Gumbel Distribution is incorrectly described
The version currently in PYTensor is the Gumbel Right Skew which gives the extreme maximum. Anyone using this to find the extreme minimum--in survival analysis for example, is using the wrong distribution. A sign change on the x and y inputs is an effective solution if the user is notified before hand.