Apply rewrites to logp before apply grad
This adds safe rewrites to logp before the grad operator is applied.
This is motivated by #6717, where expensive `cholesky(L.dot(L.T))`
operations are removed. If these remain in the logp graph when the grad
is taken, the resulting dlogp graph contains unnecessary operations.
However this may improve the stability and performance of grad logp
in other situation.