Undid test_math edits, test_continuous change
This commit undoes the changes to test_math.py by removing the
FutureWarning warnings for the _numpy functions. Note that when these
tests are removed the _numpy functions are imported explicitly in this
file. Replaced the log1mexp_numpy use in test_continuous.py by just
substituting in the maths from the function to avoid calling the
function.