Log progress via logger when no progress bar
Output terminals that do not support cursor movement ANSI control sequences
do not support rich ProgressBars, and thus need progressbar=false. However,
it is still very worthy to know how the SMC tasks are progressing.

Therefore, when progress bars are deactivated, any chance to the
state of the chains is logged to the default logger, using the exact
same state dictionary as used for the progressbars. The fact we only
log when a change happens means we do not flood the log.

The use of the logging facility also allows the called to add timestamps
in the logging handlers for a sense of elapsing time.